refactor bot handlers
