+++ ai_dietolog/bot/telegram_bot.py
@@ -1,8 +1,8 @@
 """Entry point for the AI Dietolog Telegram bot.
 
 This module initialises the Telegram application, registers basic
-commands and starts polling.  Only minimal scaffolding is provided here
-to demonstrate how the bot could be structured.
+commands and starts polling.  Specialized logic for profile setup,
+meal logging and daily review lives in ``ai_dietolog.bot.handlers``.
 
 To use this bot, set the environment variables ``TELEGRAM_BOT_TOKEN`` and
 ``OPENAI_API_KEY`` or populate ``config.json`` accordingly.  The
@@ -12,22 +12,12 @@ provided.
 
 from __future__ import annotations
 
-import json
 import logging
 import os
-from datetime import datetime
 
 from colorama import Fore, Style
 from colorama import init as colorama_init
-from openai import AsyncOpenAI  # noqa: F401
-from telegram import (
-    InlineKeyboardButton,
-    InlineKeyboardMarkup,
-    ReplyKeyboardMarkup,
-    ReplyKeyboardRemove,
-    Update,
-)
-from telegram.error import TimedOut
+from telegram import InlineKeyboardButton, InlineKeyboardMarkup, Update
 from telegram.ext import (
     Application,
     CallbackQueryHandler,
@@ -38,65 +28,11 @@ from telegram.ext import (
     filters,
 )
 
-from ..agents import profile_editor
-from ..agents.contextual import analyze_context
-from ..agents.daily_review import analyze_day as analyze_day_summary
-from ..agents.intake import intake
-from ..agents.meal_editor import edit_meal
-from ..agents.profile_collector import build_profile
-from ..core import storage
-from ..core.config import agent_llm, load_config
-from ..core.llm import ask_llm, check_llm_connectivity
-from ..core.prompts import (
-    EXTRACT_FIELD_ACTIVITY,
-    EXTRACT_FIELD_NUMERIC,
-    EXTRACT_BASIC,
-    EXTRACT_OPTIONAL,
-    AI_EXPLAIN,
-)
-from ..core.schema import (
-    Counters,
-    HistoryMeal,
-    HistoryMealEntry,
-    Meal,
-    MealBrief,
-    Norms,
-    Profile,
-    Today,
-    Total,
-)
+from .handlers import daily_review, meal_logging, profile_setup
+from ..core.config import load_config
+from ..core.llm import check_llm_connectivity
 
-logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
-
-comment_conv: ConversationHandler | None = None
-
-
-# Conversation states for conversations
-(
-    MANDATORY,
-    OPTIONAL,
-    CONFIRM,
-    EDIT,
-    MEAL_TYPE,
-    MEAL_DESC,
-    SET_PERCENT,
-    SET_COMMENT,
-    SET_CLARIFICATION,
-) = range(9)
-
-# Order of mandatory profile questions for step-by-step filling
-MANDATORY_ORDER = [
-    ("height_cm", "Укажите рост в сантиметрах:"),
-    ("weight_kg", "Ваш текущий вес (кг):"),
-    ("age", "Возраст:"),
-    ("target_weight_kg", "Желаемый вес (кг):"),
-    (
-        "activity_level",
-        "Уровень активности (sedentary/moderate/high):",
-    ),
-    ("timeframe_days", "За сколько дней хотите достичь цели?"),
-]
 
 
 async def start(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
@@ -117,830 +53,11 @@ async def handle_button_click(
     """Handle inline keyboard button clicks."""
     query = update.callback_query
     if query.data == "setup_profile":
-        return await setup_profile(update, context)
+        return await profile_setup.setup_profile(update, context)
     if query.data == "edit_profile":
-        return await start_edit_profile(update, context)
+        return await profile_setup.start_edit_profile(update, context)
     await query.answer()
-    return SET_COMMENT
-
-
-async def setup_profile(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
-    """Entry point for the profile setup conversation."""
-    if update.callback_query:
-        await update.callback_query.answer()
-    context.user_data.clear()
-    context.user_data["language"] = update.effective_user.language_code or "ru"
-    context.user_data["mandatory"] = {}
-    context.user_data["step"] = 0
-    await update.effective_message.reply_text(
-        "\U0001f4dd Давайте составим профиль. Ответьте на несколько вопросов."
-    )
-    await update.effective_message.reply_text(MANDATORY_ORDER[0][1])
-    return MANDATORY
-
-
-async def _extract(text: str, api_key: str, system: str) -> dict:
-    """Helper to call a language model and parse JSON."""
-    cfg = {**load_config(), "openai_api_key": api_key}
-    provider, model = agent_llm("extract", cfg)
-    messages = [
-        {"role": "system", "content": system},
-        {"role": "user", "content": text},
-    ]
-    if provider == "openai":
-        client = AsyncOpenAI(api_key=api_key)
-        resp = await client.chat.completions.create(
-            model=model,
-            messages=messages,
-            temperature=0,
-        )
-        content = resp.choices[0].message.content
-    else:
-        content = await ask_llm(
-            messages,
-            model=model,
-            provider=provider,
-            temperature=0,
-            cfg=cfg,
-        )
-    return json.loads(content)
-
-
-async def extract_field(field: str, text: str, api_key: str) -> dict:
-    """Extract a single profile field from ``text`` using OpenAI."""
-    if field == "activity_level":
-        system = EXTRACT_FIELD_ACTIVITY.render()
-    else:
-        system = EXTRACT_FIELD_NUMERIC.render(field=field)
-    return await _extract(text, api_key, system)
-
-
-async def extract_basic(text: str, api_key: str) -> dict:
-    """Parse mandatory profile fields from user text."""
-    system = EXTRACT_BASIC.render()
-    return await _extract(text, api_key, system)
-
-
-async def extract_optional(text: str, api_key: str) -> dict:
-    """Parse optional profile fields from user text."""
-    system = EXTRACT_OPTIONAL.render()
-    return await _extract(text, api_key, system)
-
-
-def summarise_profile(data: dict) -> str:
-    """Return a human-friendly summary of extracted data."""
-    lines = []
-    if data.get("gender"):
-        lines.append(f"Пол: {data['gender']}")
-    lines.extend(
-        [
-            f"Возраст: {data.get('age')}",
-            f"Рост: {data.get('height_cm')} см",
-            f"Вес: {data.get('weight_kg')} кг",
-            f"Целевой вес: {data.get('target_weight_kg')} кг",
-            f"Активность: {data.get('activity_level')}",
-            f"Срок: {data.get('timeframe_days')} дн.",
-        ]
-    )
-    if data.get("waist_cm"):
-        lines.append(f"Талия: {data['waist_cm']} см")
-    if data.get("bust_cm"):
-        lines.append(f"Грудь: {data['bust_cm']} см")
-    if data.get("hips_cm"):
-        lines.append(f"Бедра: {data['hips_cm']} см")
-    if data.get("restrictions"):
-        lines.append("Непереносимости: " + ", ".join(data["restrictions"]))
-    if data.get("preferences"):
-        lines.append("Предпочтения: " + ", ".join(data["preferences"]))
-    if data.get("medical"):
-        lines.append("Медицинские ограничения: " + ", ".join(data["medical"]))
-    return "\n".join(lines)
-
-
-def summarise_profile_obj(profile: Profile) -> str:
-    """Return a concise summary of a ``Profile`` object."""
-    p = profile.personal
-    g = profile.goals
-    lines = [
-        f"Пол: {p.get('gender')}",
-        f"Возраст: {p.get('age')}",
-        f"Рост: {p.get('height_cm')} см",
-        f"Вес: {p.get('weight_kg')} кг",
-        f"Активность: {p.get('activity_level')}",
-    ]
-    goal_type = g.get("type")
-    target = g.get("target_change_kg")
-    timeframe = g.get("timeframe_days")
-    if goal_type == "lose_weight":
-        goal_desc = f"снижение на {abs(target)} кг за {timeframe} дн."
-    elif goal_type == "gain_weight":
-        goal_desc = f"набор {abs(target)} кг за {timeframe} дн."
-    else:
-        goal_desc = "поддержание веса"
-    lines.append("Цель: " + goal_desc)
-    if profile.restrictions:
-        lines.append("Ограничения: " + ", ".join(profile.restrictions))
-    return "\n".join(lines)
-
-
-def meal_card(meal: Meal) -> str:
-    """Return short text describing a meal."""
-    names = ", ".join(i.name for i in meal.items)
-    t = meal.total
-    prefix = "Черновик: " if meal.pending else ""
-    return (
-        f"{prefix}{meal.type}: {names}\n"
-        f"К: {t.kcal} ккал, Б: {t.protein_g} г, Ж: {t.fat_g} г, У: {t.carbs_g} г"
-    )
-
-
-def meal_breakdown(meal: Meal) -> str:
-    """Return a detailed breakdown of a meal."""
-    lines = [f"{meal.type}:"]
-    for it in meal.items:
-        part = f"- {it.name}"
-        if it.weight_g:
-            part += f" {it.weight_g} г"
-        part += (
-            f" ({it.kcal} ккал, Б:{it.protein_g} г, Ж:{it.fat_g} г, У:{it.carbs_g} г)"
-        )
-        lines.append(part)
-    t = meal.total
-    lines.append(
-        f"Итого: {t.kcal} ккал, Б:{t.protein_g} г, Ж:{t.fat_g} г, У:{t.carbs_g} г"
-    )
-    if meal.comment:
-        lines.append(f"Комментарий: {meal.comment}")
-    prefix = "Черновик: " if meal.pending else ""
-    return prefix + "\n".join(lines)
-
-
-def format_stats(norms: Norms, summary: Total, comment: str | None = None) -> str:
-    """Return daily progress report in Markdown with emojis."""
-
-    def line(emoji: str, label: str, value: int, target: int) -> str:
-        if target:
-            percent = int(round(value / target * 100))
-            return f"{emoji} {label}: {value} / {target} ({percent}%)"
-        return f"{emoji} {label}: {value}"
-
-    lines = [
-        "\U0001f37d *Статистика дня*",
-        line("\U0001f525", "Калории", summary.kcal, norms.target_kcal),
-        line(
-            "\U0001f357", "Белки", summary.protein_g, norms.macros.get("protein_g", 0)
-        ),
-        line("\U0001f951", "Жиры", summary.fat_g, norms.macros.get("fat_g", 0)),
-        line("\U0001f35e", "Углеводы", summary.carbs_g, norms.macros.get("carbs_g", 0)),
-    ]
-    if comment:
-        lines.append("")
-        lines.append(comment)
-    return "\n".join(lines)
-
-
-async def _ai_explain(prompt: str, api_key: str) -> str:
-    """Return a short explanation from the language model."""
-    cfg = {**load_config(), "openai_api_key": api_key}
-    provider, model = agent_llm("ai_explain", cfg)
-    messages = [
-        {
-            "role": "system",
-            "content": (
-                AI_EXPLAIN.render()
-            ),
-        },
-        {"role": "user", "content": prompt},
-    ]
-    if provider == "openai":
-        client = AsyncOpenAI(api_key=api_key)
-        resp = await client.chat.completions.create(
-            model=model,
-            messages=messages,
-            temperature=0.5,
-        )
-        content = resp.choices[0].message.content
-    else:
-        content = await ask_llm(
-            messages,
-            model=model,
-            provider=provider,
-            temperature=0.5,
-            cfg=cfg,
-        )
-    return content.strip()
-
-
-async def validate_mandatory(data: dict, api_key: str) -> str | None:
-    """Return an error message if values look unrealistic.  Uses LLM for text."""
-    required = [
-        "age",
-        "height_cm",
-        "weight_kg",
-        "target_weight_kg",
-        "timeframe_days",
-        "activity_level",
-    ]
-    missing = [f for f in required if data.get(f) is None]
-    if missing:
-        mapping = {
-            "age": "возраст",
-            "height_cm": "рост",
-            "weight_kg": "вес",
-            "target_weight_kg": "целевой вес",
-            "timeframe_days": "срок",
-            "activity_level": "уровень активности",
-        }
-        human = ", ".join(mapping[m] for m in missing)
-        return await _ai_explain(f"Не удалось распознать: {human}.", api_key)
-    try:
-        age = int(data["age"])
-        height = float(data["height_cm"])
-        weight = float(data["weight_kg"])
-        target = float(data["target_weight_kg"])
-        timeframe = int(data["timeframe_days"])
-    except (TypeError, ValueError):
-        return await _ai_explain("Проверьте вводимые числа.", api_key)
-    if not 100 <= height <= 250:
-        return await _ai_explain("Рост выглядит нереалистично.", api_key)
-    if not 30 <= weight <= 300:
-        return await _ai_explain("Вес выглядит нереалистично.", api_key)
-    if not 10 <= age <= 100:
-        return await _ai_explain("Возраст выглядит нереалистично.", api_key)
-    if not 30 <= target <= 300:
-        return await _ai_explain("Целевой вес выглядит нереалистично.", api_key)
-    if timeframe <= 0:
-        return await _ai_explain("Срок должен быть больше нуля.", api_key)
-    diff = abs(weight - target)
-    max_weekly = 1.0
-    weeks = timeframe / 7
-    if weeks > 0 and diff / weeks > max_weekly:
-        needed_weeks = diff / max_weekly
-        min_days = int(needed_weeks * 7)
-        prompt = (
-            "Цель слишком быстрая. Безопасный темп — не более 1 кг в неделю. "
-            f"Для выбранной цели потребуется не менее {min_days} дней."
-        )
-        return await _ai_explain(prompt, api_key)
-    return None
-
-
-async def collect_basic(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
-    cfg = load_config()
-    api_key = cfg.get("openai_api_key") or os.getenv("OPENAI_API_KEY")
-    if not api_key:
-        await update.message.reply_text("OpenAI API key не настроен")
-        return ConversationHandler.END
-
-    step = context.user_data.get("step", 0)
-    field, _prompt = MANDATORY_ORDER[step]
-    try:
-        data = await extract_field(field, update.message.text, api_key)
-    except Exception as exc:  # noqa: BLE001
-        logger.exception("Field extraction failed: %s", exc)
-        await update.message.reply_text("Не удалось разобрать, попробуйте ещё раз")
-        return MANDATORY
-
-    value = data.get(field)
-    if value is None:
-        await update.message.reply_text("Не удалось распознать ответ, повторите")
-        return MANDATORY
-
-    context.user_data["mandatory"][field] = value
-    step += 1
-    if step < len(MANDATORY_ORDER):
-        context.user_data["step"] = step
-        await update.message.reply_text(MANDATORY_ORDER[step][1])
-        return MANDATORY
-
-    # all fields collected
-    error = await validate_mandatory(context.user_data["mandatory"], api_key)
-    if error:
-        await update.message.reply_text(error + " Попробуйте заново")
-        context.user_data["mandatory"] = {}
-        context.user_data["step"] = 0
-        await update.message.reply_text(MANDATORY_ORDER[0][1])
-        return MANDATORY
-
-    await update.message.reply_text(
-        "\U0001f4dd Теперь можете указать доп. информацию: пол, окружности, предпочтения или аллергию. Если ничего добавлять не хотите, напишите 'нет'."
-    )
-    return OPTIONAL
-
-
-async def collect_optional(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
-    cfg = load_config()
-    api_key = cfg.get("openai_api_key") or os.getenv("OPENAI_API_KEY")
-    if not api_key:
-        await update.message.reply_text("OpenAI API key не настроен")
-        return ConversationHandler.END
-    text = update.message.text.strip()
-    if text.lower() == "нет":
-        data = {}
-    else:
-        try:
-            data = await extract_optional(text, api_key)
-        except Exception as exc:  # noqa: BLE001
-            logger.exception("Optional extraction failed: %s", exc)
-            await update.message.reply_text(
-                "Не удалось разобрать сообщение, попробуйте ещё раз"
-            )
-            return OPTIONAL
-    context.user_data["optional"] = data
-    merged = {**context.user_data.get("mandatory", {}), **data}
-    summary = summarise_profile(merged)
-    context.user_data["profile"] = merged
-    await update.message.reply_text(summary + "\nВсе верно? (да/нет)")
-    return CONFIRM
-
-
-async def finish_profile(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
-    if update.message.text.strip().lower().startswith("д"):
-        data = context.user_data.get("profile", {})
-        gender = data.get("gender") or "female"
-        target_change = float(data["weight_kg"]) - float(data["target_weight_kg"])
-        goal_type = "maintain"
-        if target_change > 0:
-            goal_type = "lose_weight"
-        elif target_change < 0:
-            goal_type = "gain_weight"
-        cfg = load_config()
-        profile = await build_profile(
-            gender=gender,
-            age=int(data["age"]),
-            height_cm=float(data["height_cm"]),
-            weight_kg=float(data["weight_kg"]),
-            activity_level=data["activity_level"],
-            goal_type=goal_type,
-            target_change_kg=target_change,
-            timeframe_days=int(data["timeframe_days"]),
-            restrictions=data.get("restrictions"),
-            preferences=data.get("preferences"),
-            medical=data.get("medical"),
-            cfg=cfg,
-            language=context.user_data.get("language", "ru"),
-        )
-        for key in ("waist_cm", "bust_cm", "hips_cm"):
-            if data.get(key) is not None:
-                profile.personal[key] = data[key]
-        storage.save_profile(update.effective_user.id, profile)
-        await update.message.reply_text(
-            f"Профиль создан. Целевая калорийность: {profile.norms.target_kcal} ккал."
-        )
-        return ConversationHandler.END
-    await update.message.reply_text("Опишите данные заново:")
-    return MANDATORY
-
-
-async def cancel(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
-    await update.effective_message.reply_text("Создание профиля отменено.")
     return ConversationHandler.END
-
-
-def _end_comment_conv(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
-    """Terminate the comment conversation if it's active."""
-    global comment_conv
-    if comment_conv is None:
-        return
-    try:
-        key = comment_conv._get_key(update)
-        if key in comment_conv._conversations:
-            comment_conv._update_state(ConversationHandler.END, key)
-    except Exception:  # noqa: BLE001
-        pass
-    context.user_data.pop("comment_meal_id", None)
-    context.user_data.pop("comment_message", None)
-
-
-async def show_profile(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
-    """Send the current profile summary with an edit button."""
-    profile = storage.load_profile(update.effective_user.id, Profile)
-    if not profile.personal:
-        await update.message.reply_text(
-            "Профиль не найден. Используйте /setup_profile."
-        )
-        return ConversationHandler.END
-    text = summarise_profile_obj(profile)
-    keyboard = InlineKeyboardMarkup(
-        [
-            [
-                InlineKeyboardButton(
-                    "\u270f\ufe0f Изменить профиль", callback_data="edit_profile"
-                )
-            ]
-        ]
-    )
-    await update.message.reply_text(text, reply_markup=keyboard)
-    return ConversationHandler.END
-
-
-async def start_edit_profile(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
-    """Ask the user to describe profile changes."""
-    await update.callback_query.answer()
-    await update.callback_query.message.reply_text("Опишите изменения в профиле:")
-    return EDIT
-
-
-async def apply_profile_edit(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
-    """Update the profile via LLM and save it."""
-    cfg = load_config()
-    api_key = cfg.get("openai_api_key") or os.getenv("OPENAI_API_KEY")
-    if not api_key:
-        await update.message.reply_text("OpenAI API key не настроен")
-        return ConversationHandler.END
-    profile = storage.load_profile(update.effective_user.id, Profile)
-    try:
-        updated_dict = await profile_editor.update_profile(
-            profile.model_dump(),
-            update.message.text,
-            api_key,
-            language=context.user_data.get("language", "ru"),
-        )
-    except Exception as exc:  # noqa: BLE001
-        logger.exception("Profile update failed: %s", exc)
-        await update.message.reply_text("Не удалось обработать изменения.")
-        return ConversationHandler.END
-    try:
-        new_profile = Profile.parse_obj(updated_dict)
-    except Exception as exc:  # noqa: BLE001
-        logger.exception("Profile validation failed: %s", exc)
-        await update.message.reply_text("Получены некорректные данные.")
-        return ConversationHandler.END
-    storage.save_profile(update.effective_user.id, new_profile)
-    macros = new_profile.norms.macros
-    await update.message.reply_text(
-        "Профиль обновлён. "
-        f"Целевая калорийность {new_profile.norms.target_kcal} ккал.\n"
-        f"Б: {macros['protein_g']} г, Ж: {macros['fat_g']} г, У: {macros['carbs_g']} г"
-    )
-    return ConversationHandler.END
-
-
-async def add_meal(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
-    """Start meal logging by asking for meal type."""
-    _end_comment_conv(update, context)
-    context.user_data["language"] = update.effective_user.language_code or "ru"
-    keyboard = ReplyKeyboardMarkup(
-        [["Завтрак", "Обед"], ["Ужин", "Перекус"]],
-        one_time_keyboard=True,
-        resize_keyboard=True,
-    )
-    await update.message.reply_text("Выберите тип приёма пищи:", reply_markup=keyboard)
-    return MEAL_TYPE
-
-
-async def receive_meal_type(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
-    context.user_data["meal_type"] = update.message.text.strip()
-    history = context.user_data.setdefault("history", [])
-    history.append(update.message.text.strip())
-    del history[:-20]
-    await update.message.reply_text(
-        "Пришлите фото, голосовое или текстовое описание блюда.",
-        reply_markup=ReplyKeyboardRemove(),
-    )
-    return MEAL_DESC
-
-
-async def receive_meal_desc(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
-    """Process a meal description and photo sent by the user."""
-    desc = update.message.caption or update.message.text or ""
-    history = context.user_data.setdefault("history", [])
-    if desc:
-        history.append(desc)
-        del history[:-20]
-    image_bytes = None
-    file_id = None
-    if update.message.photo:
-        photo = update.message.photo[-1]
-        try:
-            file = await photo.get_file()
-            image_bytes = await file.download_as_bytearray()
-            file_id = photo.file_id
-        except TimedOut:
-            await update.message.reply_text(
-                "Не удалось загрузить фото, попробуйте ещё раз."
-            )
-            return MEAL_DESC
-    meal_type = context.user_data.get("meal_type", "Перекус")
-    meal = await intake(
-        image_bytes,
-        desc,
-        meal_type,
-        language=context.user_data.get("language", "ru"),
-        history=context.user_data.get("history"),
-    )
-    meal.user_desc = desc
-    meal.image_file_id = file_id
-    storage.append_meal(update.effective_user.id, meal)
-    keyboard = InlineKeyboardMarkup(
-        [
-            [
-                InlineKeyboardButton(
-                    "✔ Подтвердить", callback_data=f"confirm:{meal.id}"
-                ),
-                InlineKeyboardButton(
-                    "✍️ Комментарии", callback_data=f"comment:{meal.id}"
-                ),
-                InlineKeyboardButton("🗑 Удалить", callback_data=f"delete:{meal.id}"),
-            ]
-        ]
-    )
-    text = meal_breakdown(meal)
-    if meal.clarification:
-        text += f"\n\n❓ {meal.clarification}"
-    if file_id:
-        await update.message.reply_photo(
-            photo=file_id, caption=text, reply_markup=keyboard
-        )
-    else:
-        await update.message.reply_text(text, reply_markup=keyboard)
-
-    return ConversationHandler.END
-
-
-async def confirm_meal(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
-    query = update.callback_query
-    await query.answer()
-    _end_comment_conv(update, context)
-    meal_id = query.data.split(":", 1)[1]
-    today = storage.load_today(update.effective_user.id)
-    meal = next((m for m in today.meals if m.id == meal_id), None)
-    if not meal:
-        await query.message.reply_text("Запись не найдена")
-        return
-    if not meal.pending:
-        await query.message.reply_text("Уже подтверждено")
-        return
-    profile = storage.load_profile(update.effective_user.id, Profile)
-    cfg = load_config()
-    result = await analyze_context(
-        profile.norms.model_dump(),
-        today.summary,
-        meal.total,
-        cfg,
-        language=context.user_data.get("language", "ru"),
-        history=context.user_data.get("history"),
-    )
-    meal.pending = False
-    today.summary = Total(**result.get("summary", {}))
-    comment = result.get("context_comment")
-    if comment:
-        meal.comment = f"{meal.comment or ''} {comment}".strip()
-        history = context.user_data.setdefault("history", [])
-        history.append(comment)
-        del history[:-20]
-    storage.save_today(update.effective_user.id, today)
-    if query.message.photo:
-        await query.message.edit_caption(meal_card(meal))
-    else:
-        await query.message.edit_text(meal_card(meal))
-    stats = format_stats(profile.norms, today.summary, comment)
-    await query.message.reply_text(stats, parse_mode="Markdown")
-
-
-async def start_edit_meal(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
-    query = update.callback_query
-    await query.answer()
-    _end_comment_conv(update, context)
-    context.user_data["edit_meal_id"] = query.data.split(":", 1)[1]
-    await query.message.reply_text("Введите процент съеденного (1-100):")
-    return SET_PERCENT
-
-
-def _scale_total(total: Total, factor: float) -> Total:
-    data = {k: int(round(getattr(total, k) * factor)) for k in total.model_fields}
-    return Total(**data)
-
-
-async def apply_percent(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
-    user_id = update.effective_user.id
-    try:
-        percent = int(update.message.text)
-    except ValueError:
-        await update.message.reply_text("Введите число от 1 до 100")
-        return SET_PERCENT
-    if not 1 <= percent <= 100:
-        await update.message.reply_text("Введите число от 1 до 100")
-        return SET_PERCENT
-    meal_id = context.user_data.get("edit_meal_id")
-    today = storage.load_today(user_id)
-    meal = next((m for m in today.meals if m.id == meal_id), None)
-    if not meal:
-        await update.message.reply_text("Запись не найдена")
-        return ConversationHandler.END
-    factor = percent / meal.percent_eaten
-    old_total = meal.total
-    meal.items = [i.scale(factor) for i in meal.items]
-    meal.total = _scale_total(meal.total, factor)
-    meal.percent_eaten = percent
-    if not meal.pending:
-        for field in today.summary.model_fields:
-            value = (
-                getattr(today.summary, field)
-                - getattr(old_total, field)
-                + getattr(meal.total, field)
-            )
-            setattr(today.summary, field, value)
-    storage.save_today(user_id, today)
-    await update.message.reply_text("Изменено")
-    return ConversationHandler.END
-
-
-async def start_comment_meal(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
-    query = update.callback_query
-    await query.answer()
-    _end_comment_conv(update, context)
-    context.user_data["comment_meal_id"] = query.data.split(":", 1)[1]
-    context.user_data["comment_message"] = (
-        query.message.chat_id,
-        query.message.message_id,
-    )
-    await query.message.reply_text("Напишите комментарий к блюду:")
-    return SET_COMMENT
-
-
-async def apply_comment(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
-    user_id = update.effective_user.id
-    meal_id = context.user_data.get("comment_meal_id")
-    comment = update.message.text.strip()
-    history = context.user_data.setdefault("history", [])
-    if comment:
-        history.append(comment)
-        del history[:-20]
-    today = storage.load_today(user_id)
-    meal = next((m for m in today.meals if m.id == meal_id), None)
-    if not meal:
-        await update.message.reply_text("Запись не найдена")
-        return ConversationHandler.END
-    old_total = meal.total
-    meal.comment = f"{meal.comment or ''} {comment}".strip()
-    user_desc = f"{meal.user_desc} {comment}".strip()
-    # Refine the meal based on the comment without re-uploading the image.
-    updated = await edit_meal(
-        meal,
-        comment,
-        language=context.user_data.get("language", "ru"),
-        history=context.user_data.get("history"),
-    )
-    meal.user_desc = user_desc
-    meal.clarification = updated.clarification
-    meal.items = updated.items
-    meal.total = updated.total
-    if not meal.pending:
-        for field in today.summary.model_fields:
-            value = (
-                getattr(today.summary, field)
-                - getattr(old_total, field)
-                + getattr(meal.total, field)
-            )
-            setattr(today.summary, field, value)
-    storage.save_today(user_id, today)
-    keyboard = InlineKeyboardMarkup(
-        [
-            [
-                InlineKeyboardButton(
-                    "✔ Подтвердить", callback_data=f"confirm:{meal.id}"
-                ),
-                InlineKeyboardButton(
-                    "✍\ufe0f Комментарии", callback_data=f"comment:{meal.id}"
-                ),
-                InlineKeyboardButton("🗑 Удалить", callback_data=f"delete:{meal.id}"),
-            ]
-        ]
-    )
-    text = meal_breakdown(meal)
-    if meal.clarification:
-        text += f"\n\n❓ {meal.clarification}"
-    chat_id, msg_id = context.user_data.get(
-        "comment_message",
-        (update.effective_chat.id, update.effective_message.message_id),
-    )
-    if meal.image_file_id:
-        await context.bot.edit_message_caption(
-            chat_id=chat_id, message_id=msg_id, caption=text, reply_markup=keyboard
-        )
-    else:
-        await context.bot.edit_message_text(
-            chat_id=chat_id, message_id=msg_id, text=text, reply_markup=keyboard
-        )
-    _end_comment_conv(update, context)
-    return ConversationHandler.END
-
-
-async def delete_meal(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
-    query = update.callback_query
-    await query.answer()
-    _end_comment_conv(update, context)
-    meal_id = query.data.split(":", 1)[1]
-    user_id = update.effective_user.id
-    today = storage.load_today(user_id)
-    meal = next((m for m in today.meals if m.id == meal_id), None)
-    if not meal:
-        await query.message.reply_text("Запись не найдена")
-        return
-    if not meal.pending:
-        for field in today.summary.model_fields:
-            setattr(
-                today.summary,
-                field,
-                getattr(today.summary, field) - getattr(meal.total, field),
-            )
-    today.meals = [m for m in today.meals if m.id != meal_id]
-    storage.save_today(user_id, today)
-    if query.message.photo:
-        await query.message.edit_caption("Удалено")
-    else:
-        await query.message.edit_text("Удалено")
-
-
-async def finish_day(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
-    """Summarise the day and ask to start a new one."""
-    user_id = update.effective_user.id
-    today = storage.load_today(user_id)
-    confirmed = [m for m in today.meals if not m.pending]
-    if not confirmed:
-        await update.message.reply_text("Нет подтверждённых приёмов пищи")
-        return
-    profile = storage.load_profile(user_id, Profile)
-    cfg = load_config()
-    meal_lines = []
-    briefs = []
-    for idx, meal in enumerate(confirmed, 1):
-        t = meal.total
-        meal_lines.append(
-            f"{idx}. *{meal.type}* - {t.kcal} ккал, Б:{t.protein_g} г, "
-            f"Ж:{t.fat_g} г, У:{t.carbs_g} г"
-        )
-        meal_name = (
-            ", ".join(it.name for it in meal.items) if meal.items else meal.user_desc
-        )
-        briefs.append(
-            MealBrief(
-                type=meal.type,
-                name=meal_name,
-                **t.model_dump(),
-            )
-        )
-    stats = format_stats(profile.norms, today.summary)
-    try:
-        comment_text = await analyze_day_summary(
-            profile.norms.model_dump(),
-            today.summary,
-            briefs,
-            cfg,
-            language=context.user_data.get("language", "ru"),
-        )
-    except Exception as exc:  # noqa: BLE001
-        logger.exception("Day analysis failed: %s", exc)
-
-    text = "\U0001f4c5 *Итоги дня*\n" + "\n".join(meal_lines) + "\n\n" + stats
-    if comment_text:
-        text += "\n\n" + comment_text
-    await update.message.reply_text(text, parse_mode="Markdown")
-
-    entry = HistoryMealEntry(
-        date=datetime.utcnow().date().isoformat(),
-        num_meals=len(confirmed),
-        meals=briefs,
-        comment=comment_text,
-    )
-    context.user_data["history_entry"] = entry
-    keyboard = InlineKeyboardMarkup(
-        [
-            [
-                InlineKeyboardButton("Да", callback_data="finish_yes"),
-                InlineKeyboardButton("Нет", callback_data="finish_no"),
-            ]
-        ]
-    )
-    await update.message.reply_text("Начать новый день?", reply_markup=keyboard)
-
-
-async def confirm_finish_day(
-    update: Update, context: ContextTypes.DEFAULT_TYPE
-) -> None:
-    """Handle confirmation to start a new day."""
-    query = update.callback_query
-    await query.answer()
-    user_id = update.effective_user.id
-    if query.data == "finish_yes":
-        entry: HistoryMealEntry | None = context.user_data.get("history_entry")
-        if entry:
-            history = storage.read_json(
-                storage.json_path(user_id, "history_meal.json"), HistoryMeal
-            )
-            history.append_day(entry, max_days=60)
-            storage.write_json(storage.json_path(user_id, "history_meal.json"), history)
-        counters = storage.read_json(
-            storage.json_path(user_id, "counters.json"), Counters
-        )
-        counters.total_days_closed += 1
-        storage.write_json(storage.json_path(user_id, "counters.json"), counters)
-        storage.save_today(user_id, Today())
-        await query.message.edit_text("День завершён. Начинаем новый!")
-    else:
-        await query.message.edit_text("Отменено")
-    context.user_data.pop("history_entry", None)
 
 
 async def handle_error(update: object, context: ContextTypes.DEFAULT_TYPE) -> None:
@@ -950,7 +67,7 @@ async def handle_error(update: object, context: ContextTypes.DEFAULT_TYPE) -> No
     api_key = cfg.get("openai_api_key") or os.getenv("OPENAI_API_KEY")
     if isinstance(update, Update) and update.effective_message and api_key:
         try:
-            msg = await _ai_explain("Произошла техническая ошибка", api_key)
+            msg = await profile_setup.ai_explain("Произошла техническая ошибка", api_key)
             await update.effective_message.reply_text(msg)
         except Exception as exc:  # noqa: BLE001
             logger.exception("Failed to send error message: %s", exc)
@@ -980,68 +97,70 @@ def main() -> None:
         logger.warning("TELEGRAM_BOT_TOKEN is not set; bot will not start.")
         return
     application = Application.builder().token(token).build()
+
     conv_handler = ConversationHandler(
         entry_points=[
-            CommandHandler("setup_profile", setup_profile),
-            CallbackQueryHandler(setup_profile, pattern="^setup_profile$"),
+            CommandHandler("setup_profile", profile_setup.setup_profile),
+            CallbackQueryHandler(profile_setup.setup_profile, pattern="^setup_profile$"),
         ],
         states={
-            MANDATORY: [MessageHandler(filters.TEXT & ~filters.COMMAND, collect_basic)],
-            OPTIONAL: [
-                MessageHandler(filters.TEXT & ~filters.COMMAND, collect_optional)
+            profile_setup.MANDATORY: [
+                MessageHandler(filters.TEXT & ~filters.COMMAND, profile_setup.collect_basic)
             ],
-            CONFIRM: [MessageHandler(filters.TEXT & ~filters.COMMAND, finish_profile)],
+            profile_setup.OPTIONAL: [
+                MessageHandler(filters.TEXT & ~filters.COMMAND, profile_setup.collect_optional)
+            ],
+            profile_setup.CONFIRM: [
+                MessageHandler(filters.TEXT & ~filters.COMMAND, profile_setup.finish_profile)
+            ],
         },
-        fallbacks=[CommandHandler("cancel", cancel)],
+        fallbacks=[CommandHandler("cancel", profile_setup.cancel)],
     )
 
     edit_conv = ConversationHandler(
-        entry_points=[
-            CallbackQueryHandler(start_edit_profile, pattern="^edit_profile$")
-        ],
+        entry_points=[CallbackQueryHandler(profile_setup.start_edit_profile, pattern="^edit_profile$")],
         states={
-            EDIT: [MessageHandler(filters.TEXT & ~filters.COMMAND, apply_profile_edit)],
+            profile_setup.EDIT: [
+                MessageHandler(filters.TEXT & ~filters.COMMAND, profile_setup.apply_profile_edit)
+            ],
         },
-        fallbacks=[CommandHandler("cancel", cancel)],
+        fallbacks=[CommandHandler("cancel", profile_setup.cancel)],
     )
 
     meal_conv = ConversationHandler(
-        entry_points=[CommandHandler("add_meal", add_meal)],
+        entry_points=[CommandHandler("add_meal", meal_logging.add_meal)],
         states={
-            MEAL_TYPE: [
-                MessageHandler(filters.TEXT & ~filters.COMMAND, receive_meal_type)
+            meal_logging.MEAL_TYPE: [
+                MessageHandler(filters.TEXT & ~filters.COMMAND, meal_logging.receive_meal_type)
             ],
-            MEAL_DESC: [
-                MessageHandler(
-                    (filters.TEXT | filters.PHOTO) & ~filters.COMMAND, receive_meal_desc
-                )
+            meal_logging.MEAL_DESC: [
+                MessageHandler((filters.TEXT | filters.PHOTO) & ~filters.COMMAND, meal_logging.receive_meal_desc)
             ],
-            SET_COMMENT: [
-                MessageHandler(filters.TEXT & ~filters.COMMAND, apply_comment)
+            meal_logging.SET_COMMENT: [
+                MessageHandler(filters.TEXT & ~filters.COMMAND, meal_logging.apply_comment)
             ],
         },
-        fallbacks=[CommandHandler("cancel", cancel)],
+        fallbacks=[CommandHandler("cancel", profile_setup.cancel)],
     )
 
     edit_meal_conv = ConversationHandler(
-        entry_points=[CallbackQueryHandler(start_edit_meal, pattern="^edit:")],
+        entry_points=[CallbackQueryHandler(meal_logging.start_edit_meal, pattern="^edit:")],
         states={
-            SET_PERCENT: [
-                MessageHandler(filters.TEXT & ~filters.COMMAND, apply_percent)
+            meal_logging.SET_PERCENT: [
+                MessageHandler(filters.TEXT & ~filters.COMMAND, meal_logging.apply_percent)
             ],
         },
-        fallbacks=[CommandHandler("cancel", cancel)],
+        fallbacks=[CommandHandler("cancel", profile_setup.cancel)],
     )
 
-    global comment_conv
-    comment_conv = ConversationHandler(
-        entry_points=[CallbackQueryHandler(start_comment_meal, pattern="^comment:")],
+    meal_logging.comment_conv = ConversationHandler(
+        entry_points=[CallbackQueryHandler(meal_logging.start_comment_meal, pattern="^comment:")],
         states={
-            SET_COMMENT: [
-                MessageHandler(filters.TEXT & ~filters.COMMAND, apply_comment)
+            meal_logging.SET_COMMENT: [
+                MessageHandler(filters.TEXT & ~filters.COMMAND, meal_logging.apply_comment)
             ],
         },
-        fallbacks=[CommandHandler("cancel", cancel)],
+        fallbacks=[CommandHandler("cancel", profile_setup.cancel)],
         allow_reentry=True,
     )
 
@@ -1049,14 +168,14 @@ def main() -> None:
     application.add_handler(edit_conv)
     application.add_handler(meal_conv)
     application.add_handler(edit_meal_conv)
-    application.add_handler(comment_conv)
+    application.add_handler(meal_logging.comment_conv)
     application.add_handler(CommandHandler("start", start))
-    application.add_handler(CommandHandler("profile", show_profile))
-    application.add_handler(CommandHandler("finish_day", finish_day))
-    application.add_handler(CallbackQueryHandler(confirm_meal, pattern="^confirm:"))
-    application.add_handler(CallbackQueryHandler(delete_meal, pattern="^delete:"))
+    application.add_handler(CommandHandler("profile", profile_setup.show_profile))
+    application.add_handler(CommandHandler("finish_day", daily_review.finish_day))
+    application.add_handler(CallbackQueryHandler(meal_logging.confirm_meal, pattern="^confirm:"))
+    application.add_handler(CallbackQueryHandler(meal_logging.delete_meal, pattern="^delete:"))
     application.add_handler(
-        CallbackQueryHandler(confirm_finish_day, pattern="^finish_(yes|no)$")
+        CallbackQueryHandler(daily_review.confirm_finish_day, pattern="^finish_(yes|no)$")
     )
     application.add_handler(CallbackQueryHandler(handle_button_click))
     application.add_error_handler(handle_error)
