+++ docs/MODULES.md
@@ -12,7 +12,10 @@
 
 Код взаимодействия с Telegram:
 
-- **telegram_bot.py** – точка входа для бота. Инициализирует приложение, регистрирует команды, обрабатывает диалог сбора профиля. Для разбора текста обращается к OpenAI и использует `agents.profile_collector` для построения профиля.
+- **telegram_bot.py** – точка входа для бота. Инициализирует приложение и регистрирует обработчики.
+- **handlers/profile_setup.py** – заполнение и редактирование профиля.
+- **handlers/meal_logging.py** – добавление, редактирование и форматирование приёмов пищи.
+- **handlers/daily_review.py** – суточные отчёты и подсчёт статистики.
 
 ## ai_dietolog/core
